+++ CHANGELOG.md
@@ -19,3 +19,5 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add password reset flow - #147 by @dominik-zeglen
 - Add support for multiple values in filters - #160 by @dominik-zeglen
 - UI improvements - #166 by @benekex2
+- Fix en locale matching - #165 by @dominik-zeglen
+